django_openfoodfact-main/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
pymongo==3.12.0
pytz==2024.1
requests==2.32.3
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.2
